Finished

+++ src/module1/_3expressions/E05OperatorExample.java
@@ -8,7 +8,7 @@ public class E05OperatorExample {
 		System.out.println(2 + 3);
 		System.out.println(2 - 3);
 		System.out.println(2 * 3);
-		System.out.println(2 / 3);
+		System.out.println((double) 2 / 3);
 		System.out.println(2 == 3);
 		System.out.println(2 != 3);
 	}
